backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-RESTX==1.3.0

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.7
alembic==1.13.0

# WebAuthn
webauthn==2.0.0
cbor2==5.5.1

# Security
bcrypt==4.1.2
python-dotenv==1.0.0
werkzeug==3.0.1

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
